sprites(kinda), cpu quirk and fixes
made basic sprites handling, still need to figure out some things about the rendering and most importantly the timing. the cpu interrupt quirk is to pass a test

+++ src/platform.h
@@ -20,7 +20,7 @@ public:
 	}
 
 	void mainLoop(Gameboy& gb);
-	void updateScreen(uint8* data);
+	void updateScreen(uint16* data);
 	void render() const;
 
 private:
